FFX Controller minRefresh and tweak FFXSegment::updateOverlay
Added a minRefresh timer to FFXController to force a refresh every so often (default to 1000 ms).  This is handy for static colors/effects that don't update - if the LEDS glitch or LED power is interrupted, the frame will be refreshed even if the effect is not updated.

Fixed updateOverlay in FFXSegment to draw the last frame and then check on the next cycle if the overlay is complete.

Fixed zipOverlayFX to clear the last frame at the end of each cycle